serverless/search/requirements.txt
aiohttp==3.8.3
  - aiosignal [required: >=1.1.2, installed: 1.2.0]
    - frozenlist [required: >=1.1.0, installed: 1.3.1]
  - async-timeout [required: <5.0,>=4.0.0a3, installed: 4.0.2]
  - attrs [required: >=17.3.0, installed: 22.1.0]
  - charset-normalizer [required: >=2.0,<3.0, installed: 2.1.1]
  - frozenlist [required: >=1.1.1, installed: 1.3.1]
  - multidict [required: <7.0,>=4.5, installed: 6.0.2]
  - yarl [required: >=1.0,<2.0, installed: 1.8.1]
    - idna [required: >=2.0, installed: 3.4]
    - multidict [required: >=4.0, installed: 6.0.2]
bs4==0.0.1
  - beautifulsoup4 [required: Any, installed: 4.11.1]
    - soupsieve [required: >1.2, installed: 2.3.2.post1]
elasticsearch==8.4.3
  - elastic-transport [required: >=8,<9, installed: 8.4.0]
    - certifi [required: Any, installed: 2022.9.24]
    - urllib3 [required: >=1.26.2,<2, installed: 1.26.12]
fastapi==0.85.1
  - pydantic [required: !=1.8,!=1.7.3,!=1.7.2,!=1.7.1,<2.0.0,>=1.6.2,!=1.7,!=1.8.1, installed: 1.10.2]
    - typing-extensions [required: >=4.1.0, installed: 4.4.0]
  - starlette [required: ==0.20.4, installed: 0.20.4]
    - anyio [required: <5,>=3.4.0, installed: 3.6.1]
      - idna [required: >=2.8, installed: 3.4]
      - sniffio [required: >=1.1, installed: 1.3.0]
    - typing-extensions [required: >=3.10.0, installed: 4.4.0]
lxml==4.9.1
python-dotenv==0.21.0

